Move core/process to its own package
