+++ .vscode/settings.json
@@ -23,6 +23,7 @@
         "packages/navigator/coverage/lcov.info",
         "packages/preferences-api/coverage/lcov.info",
         "packages/preferences/coverage/lcov.info",
+        "packages/process/coverage/lcov.info",
         "packages/python/coverage/lcov.info",
         "packages/terminal/coverage/lcov.info",
         "packages/workspace/coverage/lcov.info"
