+++ examples/browser/.yo-rc.json
@@ -9,6 +9,7 @@
       "@theia/workspace": "../../packages/workspace",
       "@theia/preferences": "../../packages/preferences",
       "@theia/navigator": "../../packages/navigator",
+      "@theia/process": "../../packages/process",
       "@theia/terminal": "../../packages/terminal",
       "@theia/editor": "../../packages/editor",
       "@theia/languages": "../../packages/languages",
