+++ examples/electron/theia.package.json
@@ -12,6 +12,7 @@
     "@theia/languages": "^0.1.1",
     "@theia/monaco": "^0.1.1",
     "@theia/navigator": "^0.1.1",
+    "@theia/process": "^0.1.1",
     "@theia/preferences": "^0.1.1",
     "@theia/python": "^0.1.1",
     "@theia/terminal": "^0.1.1",
