+++ packages/process/README.md
@@ -1,0 +1,8 @@
+# Theia - Process Extension
+
+This extension allows the management of proccesses started with or without a terminal.
+
+See [here](https://github.com/theia-ide/theia) for a detailed documentation.
+
+## License
+[Apache-2.0](https://github.com/theia-ide/theia/blob/master/LICENSE)
