+++ packages/process/compile.tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../base.tsconfig",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "lib"
+  },
+  "include": [
+    "src"
+  ]
+}
