+++ packages/process/src/node/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright (C) 2017 Ericsson and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+export * from './process-manager';
+export * from './process';
+export * from './raw-process';
+export * from './terminal-process';
